Avoid make bug that swallows warnings
Building 'LuaTest(does-not-exist)' was not reporting an error.  It
turns out that Make 3.81 will not report an error when a target has a
non-existent preqrequisite when there is an included makefile that
does not exist, and there exists a rule to build it, triggering Make's
auto-restart feature, and the rule to build that makefile fails
because of a non-existent prerequisite, and that prerequisite is the
same file causing the command-line specified target to fail. (See
build/makebug.mk for an example.)

The LuaTest and LuaScan rules, in combination, were running afoul of
this.  Also JSTest/JSScan.

The solution is to change LuaScan to omit its input file from its list
of prerequisites, and then to have the makefile it generates declare
dependencies of the makefile itself, as well as the test result
target.

Added `-MTF` to cfromlua and jsdep to handle this case.  This might
acutally address other shortcomings, capturing dependencies of the
generated makefile that otherwise weren't.  This option should be
appropriate for all dependency makefiles that are generated with their
own rule (rather than as a side effect of generating some other
target).